move DummyClient classes into their own package
The DummyClient classes are not directly part of net.java.otr4j.session, so
it doesn't make sense for them to be there.  The DummyClient is also useful
for testing classes in other parts of otr4j, so that is another reason to
split it out into its own package.  Then only the tests that directly
relate to the classes of a given package will be included in the test
version of that package.
