+++ src/test/java/net/java/otr4j/test/dummyclient/Connection.java
@@ -1,11 +1,11 @@
-package net.java.otr4j.session;
+package net.java.otr4j.test.dummyclient;
 
 import net.java.otr4j.OtrException;
 
 /**
  * Created by gp on 2/6/14.
  */
-class Connection {
+public class Connection {
 
 	private final DummyClient client;
 	private final String connectionName;
